requirements.txt
azure-cognitiveservices-speech >= 1.20.0
requests
Deprecated